Refine self-test config and shared error responses
